Support VFR clips by using a Python vspipe --timecodes implementation
Might be extremely slow for certain source filters and longer clips!
Might become much faster when the VideoNode.frames generator is updated.

Changes to f2ts:
  - Make it a public function
  - Change signature to take frame number as first positional-only arg
  - Allow returning different precision timestamps
    * This is important for external script using this to generate
      chapter timestamps with nanosecond precision
  - Support VFR clips by using a cached "list" of timestamps per frame

Ideally, the public API should look something like this now:

You should only be trying to find frame-based timestamps on source
filter clips, and should be calling it using a functools.partial,
with the precision and src_clip keyword args defined.

If using VFR clips, expect there to be some delay as a timecode list
has to be generated using frame props as the clip's fps attribute will
be 0. As the frames need to be accessed, DON'T use this on a clip that
is being filtered.

An example usage would be

ts = partial(f2ts, precision=6, src_clip=VFR_clip)
[ts(i) for i in [8, 90, 101, 456, ...]]